+++ resources/views/account/edit.blade.php
@@ -1,9 +1,9 @@
-@extends('master')
+@extends('main')
 
 @section('content')
 
     <h2>Edit account details</h2>
-    @include('includes.page.breadcrumb', ['crumbs' => Request::segments()])
+    @include('includes.page.breadcrumb')
     <div class="panel panel-default">
         <div class="panel-heading"><h3 class="panel-title">USER DETAILS</h3></div>
         <div class="panel-body">