+++ resources/views/applications/partial-refund.blade.php
@@ -1,9 +1,9 @@
-@extends('master')
+@extends('main')
 
 @section('content')
 
     <h2>Partial Refund</h2>
-    @include('includes.page.breadcrumb', ['crumbs' => Request::segments()])
+    @include('includes.page.breadcrumb', ['over' => [1 => $application->installation->name], 'permission' => [0 => Auth::user()->can('merchants-view'), 1 => Auth::user()->can('merchants-view')]])
 
     <p>&nbsp;</p>
     {!! Form::open( ['method'=>'post', 'class' => 'form-horizontal'] ) !!}