+++ resources/views/applications/pending-cancellation.blade.php
@@ -1,8 +1,9 @@
-@extends('master')
+@extends('main')
 
 @section('content')
 
     <h2>Pending Cancellations</h2>
+    @include('includes.page.breadcrumb', ['over' => [1 => isset($application[0]->installation->name) ? $application[0]->installation->name : Request::segment(2)], 'permission' => [0 => Auth::user()->can('merchants-view'), 1 => Auth::user()->can('merchants-view')]])
 
     {!! Form::open(array('url' => Request::url() . '/?' . Request::server('QUERY_STRING'), 'method' => 'get',  'onsubmit'=>"return submitFilter()")) !!}