+++ resources/views/includes/page/breadcrumb.blade.php
@@ -1,11 +1,12 @@
 <ol class="breadcrumb">
     {{-- */$url='';/* --}}
+    {{-- */$crumbs=Request::segments();/* --}}
     <li><a href="/"{!! $crumbs[0] !!}>Dashboard</a></li>
     @foreach($crumbs as $k => $v)
 
         {{-- */$url.='/'.$v;/* --}}
         <li>
-        @if(end($crumbs) == $v)
+        @if(end($crumbs) == $v || (isset($permission[$k]) && (!$permission[$k])))
             {!! ucwords(str_replace('-', ' ', isset($over[$k])?$over[$k]:$v)) !!}
         @else
         <a href="{!!$url!!}">{!! ucwords(str_replace('-', ' ', isset($over[$k])?$over[$k]:$v)) !!}</a>