+++ resources/views/merchants/create.blade.php
@@ -1,9 +1,9 @@
-@extends('master')
+@extends('main')
 
 @section('content')
 
     <h2>{{ Str::upper(' create a new ' . Request::segment(1)) }}</h2>
-    @include('includes.page.breadcrumb', ['crumbs' => Request::segments()])
+    @include('includes.page.breadcrumb')
 
     {!! Form::open(['url' => 'merchants', 'class' => 'form-horizontal']) !!}
     <p>&nbsp;</p>